Had to change one of the parameters in the Repository interfaces to Integer instead of Long. Added userId as a attribute to grocery, card and contactinformation so that it can be saved to the correct user. Started writing the CartController.java

+++ src/main/java/se/newton/project/CrudRepositories/ContactInformationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import se.newton.project.DBModels.ContactInformation;
 
-public interface ContactInformationRepository extends CrudRepository<ContactInformation, Long> {
+public interface ContactInformationRepository extends CrudRepository<ContactInformation, Integer> {
 
 }
