+++ src/main/java/se/newton/project/CrudRepositories/GroceryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import se.newton.project.DBModels.Grocery;
 
-public interface GroceryRepository extends CrudRepository<Grocery, Long> {
+public interface GroceryRepository extends CrudRepository<Grocery, Integer> {
 
 }
